Restructure the code - Update dependencies - Can add TODO from MainActivity
