+++ app/src/main/java/com/biryanistudio/todo/services/CopyListenerService.java
@@ -1,4 +1,4 @@
-package com.biryanistudio.todo.utils;
+package com.biryanistudio.todo.services;
 
 import android.animation.Animator;
 import android.app.Service;
@@ -21,6 +21,7 @@ import android.view.animation.AccelerateDecelerateInterpolator;
 import android.widget.TextView;
 
 import com.biryanistudio.todo.R;
+import com.biryanistudio.todo.db.DbTransactions;
 import com.biryanistudio.todo.ui.MainActivity;
 
 public class CopyListenerService extends Service implements
